增删改查基本完结 接下来我这里主要是 //TODO 导出.sql

+++ back/backend/src/main/java/com/whut/backend/entity/QuestionBankQuestion.java
@@ -11,9 +11,9 @@ import java.io.Serializable;
 @TableName(value ="question_bank_question")
 @Data
 public class QuestionBankQuestion implements Serializable {
-    private Long id;
-    private Long questionBankId;
-    private Long questionId;
+    private Integer id;
+    private Integer questionBankId;
+    private Integer questionId;
 
     private static final long serialVersionUID = 1L;
 }
